backend/requirements.txt
django
djangorestframework
djongo
neomodel>=5.3.0
djangorestframework-simplejwt
python-dotenv
mongoengine
Werkzeug 
google-generativeai
django-cors-headers
google-generativeai
langchain==0.2.12
langchain-core==0.2.28
langchain-community==0.2.11
langchain-groq==0.1.9
groq
langgraph
# Force Rebuild 1